type state files: action, reducers, ReducWrapper, and store
