+++ declarations.d.ts
@@ -4,3 +4,6 @@ declare module "*.css" {
 }
 
 declare module "*.png"
+
+declare module 'redux-persist/lib/storage';
+declare module 'redux-persist/integration/react';
