+++ src/state/store.tsx
@@ -1,16 +1,16 @@
-import { createStore, applyMiddleware} from 'redux';
+import { createStore, applyMiddleware, Store} from 'redux';
 import { persistStore, persistReducer } from 'redux-persist';
 import {composeWithDevTools} from 'redux-devtools-extension';
 import { createLogger } from 'redux-logger'
 import storage from 'redux-persist/lib/storage';
-import rootReducer from './reducers';
+import appReducer from './reducers';
 
 const persistConfig = {
   key: 'root',
   storage,
 };
 
-const persistedReducer = persistReducer(persistConfig, rootReducer);
+const persistedReducer = persistReducer(persistConfig, appReducer);
 const loggerMiddleware = createLogger();
 const middleware = composeWithDevTools(applyMiddleware(loggerMiddleware));
